# backend/docs/openapi.yaml
components:
  schemas:
...
    Code:
      properties:
        code:
          maxLength: 128
          minLength: 1
          title: Code
          type: string
      required:
      - code
      title: Code
      type: object
...
    Email:
      properties:
        email:
          format: email
          maxLength: 128
          title: Email
          type: string
      required:
      - email
      title: Email
      type: object
    ErrorResponse:
      properties:
        detail:
          title: Detail
          type: string
      required:
      - detail
      title: ErrorResponse
      type: object
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
    LoginUserData:
      properties:
        email:
          title: Email
          type: string
        password:
          title: Password
          type: string
      required:
      - email
      - password
      title: LoginUserData
      type: object
...
    Password:
      properties:
        password:
          maxLength: 128
          minLength: 6
          title: Password
          type: string
      required:
      - password
      title: Password
      type: object
    RegisterUserData:
      properties:
        do_verify_email:
          title: Do Verify Email
          type: boolean
        email:
          format: email
          maxLength: 128
          title: Email
          type: string
        first_name:
          maxLength: 128
          minLength: 1
          title: First Name
          type: string
        last_name:
          maxLength: 128
          minLength: 1
          title: Last Name
          type: string
        password:
          maxLength: 128
          minLength: 6
          title: Password
          type: string
      required:
      - first_name
      - last_name
      - email
      - password
      - do_verify_email
      title: RegisterUserData
      type: object
...
    TokenInfo:
      properties:
        access_token:
          title: Access Token
          type: string
        token_type:
          default: Bearer
          title: Token Type
          type: string
      required:
      - access_token
      title: TokenInfo
      type: object
    UserSchema:
      properties:
        additional_emails:
          anyOf:
          - items: {}
            type: array
          - type: 'null'
          title: Additional Emails
        email:
          title: Email
          type: string
        first_name:
          title: First Name
          type: string
        id:
          format: uuid
          title: Id
          type: string
        is_active:
          title: Is Active
          type: boolean
        is_verified:
          title: Is Verified
          type: boolean
        last_name:
          title: Last Name
          type: string
        nickname:
          anyOf:
          - type: string
          - type: 'null'
          title: Nickname
      required:
      - id
      - first_name
      - last_name
      - nickname
      - is_active
      - is_verified
      - email
      - additional_emails
      title: UserSchema
      type: object
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
  securitySchemes:
    HTTPBearer:
      scheme: bearer
      type: http
info:
  description: "API \u0434\u043B\u044F Termeet"
  title: Termeet API
  version: 1.0.0
openapi: 3.1.0
paths:
  /auth/confirm-email:
    post:
      description: "\u041F\u043E\u0434\u0442\u0432\u0435\u0440\u0436\u0434\u0430\u0435\
        \u0442 email, \u043E\u0442\u043F\u0440\u0430\u0432\u043B\u044F\u0435\u0442\
        \ \u043F\u0438\u0441\u044C\u043C\u043E \u0441 \u043F\u043E\u0434\u0442\u0432\
        \u0435\u0440\u0436\u0434\u0435\u043D\u0438\u0435\u043C"
      operationId: confirm_email_auth_confirm_email_post
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
      security:
      - HTTPBearer: []
      summary: "\u041F\u043E\u0434\u0442\u0432\u0435\u0440\u0436\u0434\u0435\u043D\
        \u0438\u0435 email"
      tags:
      - Auth
  /auth/confirm-email/verify:
    post:
      description: "\u041F\u0440\u043E\u0432\u0435\u0440\u044F\u0435\u0442 \u0432\u0430\
        \u043B\u0438\u0434\u043D\u043E\u0441\u0442\u044C \u0442\u043E\u043A\u0435\u043D\
        \u0430 \u0432\u0435\u0440\u0438\u0444\u0438\u043A\u0430\u0446\u0438\u0438\
        \ \u043F\u043E\u0447\u0442\u044B"
      operationId: verify_token_auth_confirm_email_verify_post
      parameters:
      - in: query
        name: token
        required: true
        schema:
          title: Token
          type: string
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: "\u0421\u0440\u043E\u043A \u0434\u0435\u0439\u0441\u0442\u0432\
            \u0438\u044F \u0442\u043E\u043A\u0435\u043D\u0430 \u0438\u0441\u0442\u0435\
            \u043A                 \u041D\u0435 \u043F\u0440\u0430\u0432\u0438\u043B\
            \u044C\u043D\u044B\u0439 \u0442\u0438\u043F \u0442\u043E\u043A\u0435\u043D\
            \u0430"
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: "\u041F\u0440\u043E\u0432\u0435\u0440\u043A\u0430 \u0442\u043E\u043A\
        \u0435\u043D\u0430 \u0432\u0435\u0440\u0438\u0444\u0438\u043A\u0430\u0446\u0438\
        \u0438 \u043F\u043E\u0447\u0442\u044B"
      tags:
      - Auth
  /auth/login:
    post:
      description: "\u0410\u0432\u0442\u043E\u0440\u0438\u0437\u0443\u0435\u0442 \u043F\
        \u043E\u043B\u044C\u0437\u043E\u0432\u0430\u0442\u0435\u043B\u044F \u043F\u043E\
        \ email \u0438 \u043F\u0430\u0440\u043E\u043B\u044E,         \u0432\u044B\u0434\
        \u0430\u0451\u0442 access \u0438 refresh \u0442\u043E\u043A\u0435\u043D"
      operationId: auth_user_issue_jwt_auth_login_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginUserData'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenInfo'
          description: Successful Response
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: "\u041D\u0435 \u0432\u0435\u0440\u043D\u044B\u0439 \u043B\u043E\
            \u0433\u0438\u043D \u0438\u043B\u0438 \u043F\u0430\u0440\u043E\u043B\u044C"
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: "\u041F\u043E\u043B\u044C\u0437\u043E\u0432\u0430\u0442\u0435\
            \u043B\u044C \u0437\u0430\u0431\u043B\u043E\u043A\u0438\u0440\u043E\u0432\
            \u0430\u043D"
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: "\u0410\u0432\u0442\u043E\u0440\u0438\u0437\u0430\u0446\u0438\u044F\
        \ \u043F\u043E\u043B\u044C\u0437\u043E\u0432\u0430\u0442\u0435\u043B\u044F\
        , \u0432\u044B\u0434\u0430\u0447\u0430 \u0442\u043E\u043A\u0435\u043D\u043E\
        \u0432"
      tags:
      - Auth
...
  /auth/refresh:
    post:
      description: "\u041E\u0431\u043D\u043E\u0432\u043B\u044F\u0435\u0442\u0441\u044F\
        \ access \u0442\u043E\u043A\u0435\u043D \u043F\u043E refresh \u0442\u043E\u043A\
        \u0435\u043D\u0443"
      operationId: auth_refresh_jwt_auth_refresh_post
      parameters:
      - in: cookie
        name: refresh_token
        required: true
        schema:
          title: Refresh Token
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenInfo'
          description: Successful Response
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: "\u0421\u0440\u043E\u043A \u0434\u0435\u0439\u0441\u0442\u0432\
            \u0438\u044F refresh-\u0442\u043E\u043A\u0435\u043D\u0430 \u0438\u0441\
            \u0442\u0435\u043A                 \u041D\u0435 \u043F\u0440\u0430\u0432\
            \u0438\u043B\u044C\u043D\u044B\u0439 \u0442\u0438\u043F \u0442\u043E\u043A\
            \u0435\u043D\u0430"
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: "\u041E\u0431\u043D\u043E\u0432\u0438\u0442\u044C access \u0442\u043E\
        \u043A\u0435\u043D"
      tags:
      - Auth
  /auth/register:
    post:
      description: "\u0420\u0435\u0433\u0438\u0441\u0442\u0440\u0438\u0440\u0443\u0435\
        \u0442 \u043D\u043E\u0432\u043E\u0433\u043E \u043F\u043E\u043B\u044C\u0437\
        \u043E\u0432\u0430\u0442\u0435\u043B\u044F, \u043F\u043E\u043B\u0443\u0447\
        \u0430\u0435\u0442 \u0435\u0433\u043E \u0434\u0430\u043D\u043D\u044B\u0435\
        ,                  \u0445\u0435\u0448\u0438\u0440\u0443\u0435\u0442 \u043F\
        \u0430\u0440\u043E\u043B\u044C, \u0441\u043E\u0445\u0440\u0430\u043D\u044F\
        \u0435\u0442 \u0432 \u0411\u0414; \u0442\u0430\u043A\u0436\u0435 \u043E\u0442\
        \u043F\u0440\u0430\u0432\u043B\u044F\u0435\u0442                  \u043F\u0438\
        \u0441\u044C\u043C\u043E \u0441 \u043F\u043E\u0434\u0442\u0432\u0435\u0440\
        \u0436\u0434\u0435\u043D\u0438\u0435\u043C"
      operationId: default_register_user_auth_register_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterUserData'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserSchema'
          description: Successful Response
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: "\u041F\u043E\u043B\u044C\u0437\u043E\u0432\u0430\u0442\u0435\
            \u043B\u044C \u0441 \u044D\u0442\u0438 email-\u043E\u043C \u0443\u0436\
            \u0435 \u0441\u0443\u0449\u0435\u0441\u0442\u0432\u0443\u0435\u0442"
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: "\u0420\u0435\u0433\u0438\u0441\u0442\u0440\u0430\u0446\u0438\u044F\
        \ \u043D\u043E\u0432\u043E\u0433\u043E \u043F\u043E\u043B\u044C\u0437\u043E\
        \u0432\u0430\u0442\u0435\u043B\u044F"
      tags:
      - Auth
  /auth/reset-password:
    post:
      description: "\u041F\u043E\u0434\u0442\u0432\u0435\u0440\u0436\u0434\u0430\u0435\
        \u0442 email, \u043E\u0442\u043F\u0440\u0430\u0432\u043B\u044F\u0435\u0442\
        \ \u043F\u0438\u0441\u044C\u043C\u043E \u0441 \u043F\u043E\u0434\u0442\u0432\
        \u0435\u0440\u0436\u0434\u0435\u043D\u0438\u0435\u043C,                  \u043E\
        \u0442\u043F\u0440\u0430\u0432\u043B\u044F\u0435\u0442 \u043F\u0438\u0441\u044C\
        \u043C\u043E \u0441 \u0441\u0441\u044B\u043B\u043A\u043E\u0438\u0306 \u0434\
        \u043B\u044F \u0441\u0431\u0440\u043E\u0441\u0430 \u043F\u0430\u0440\u043E\
        \u043B\u044F"
      operationId: reset_password_auth_reset_password_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Email'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: "\u0421\u0431\u0440\u043E\u0441 \u043F\u0430\u0440\u043E\u043B\u044F"
      tags:
      - Auth
  /auth/reset-password/verify:
    post:
      description: "\u041F\u0440\u043E\u0432\u0435\u0440\u044F\u0435\u0442 \u0432\u0430\
        \u043B\u0438\u0434\u043D\u043E\u0441\u0442\u044C \u0442\u043E\u043A\u0435\u043D\
        \u0430 \u0441\u0431\u0440\u043E\u0441\u0430 \u043F\u0430\u0440\u043E\u043B\
        \u044F"
      operationId: verify_reset_password_token_auth_reset_password_verify_post
      parameters:
      - in: query
        name: token
        required: true
        schema:
          title: Token
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Password'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: "\u0421\u0440\u043E\u043A \u0434\u0435\u0439\u0441\u0442\u0432\
            \u0438\u044F \u0442\u043E\u043A\u0435\u043D\u0430 \u0438\u0441\u0442\u0435\
            \u043A                 \u041D\u0435 \u043F\u0440\u0430\u0432\u0438\u043B\
            \u044C\u043D\u044B\u0439 \u0442\u0438\u043F \u0442\u043E\u043A\u0435\u043D\
            \u0430"
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: "\u041E\u0431\u044A\u0435\u043A\u0442 \u043D\u0435 \u043D\u0430\
            \u0439\u0434\u0435\u043D"
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: "\u041F\u0440\u043E\u0432\u0435\u0440\u043A\u0430 \u0442\u043E\u043A\
        \u0435\u043D\u0430 \u0441\u0431\u0440\u043E\u0441\u0430 \u043F\u0430\u0440\
        \u043E\u043B\u044F \u0438 \u0437\u0430\u043C\u0435\u043D\u0430 \u043F\u0430\
        \u0440\u043E\u043B\u044F \u043D\u0430 \u043D\u043E\u0432\u044B\u0439"
      tags:
      - Auth
  /auth/yandex/callback:
    post:
      description: "\u041A\u043E\u0434 \u043E\u0442\u043F\u0440\u0430\u0432\u043B\u044F\
        \u0435\u0442\u0441\u044F \u043D\u0430 \u0440\u0443\u0447\u043A\u0443 \u042F\
        \u043D\u0434\u0435\u043A\u0441\u0430                 \u0434\u043B\u044F \u043F\
        \u043E\u043B\u0443\u0447\u0435\u043D\u0438\u044F \u0442\u043E\u043A\u0435\u043D\
        \u043E\u0432, \u043F\u043E\u0441\u043B\u0435 \u0443\u0436\u0435 access \u0442\
        \u043E\u043A\u0435\u043D                 \u043E\u0442\u043F\u0440\u0430\u0432\
        \u043B\u044F\u0435\u0442\u0441\u044F \u043D\u0430 \u0440\u0443\u0447\u043A\
        \u0443 \u042F\u043D\u0434\u0435\u043A\u0441\u0430, \u0434\u043B\u044F \u043F\
        \u043E\u043B\u0443\u0447\u0435\u043D\u0438\u044F                  \u0434\u043E\
        \u0441\u0442\u0443\u043F\u043D\u043E\u0439 \u0438\u043D\u0444\u043E\u0440\u043C\
        \u0430\u0446\u0438\u0438 \u043E \u043F\u043E\u043B\u044C\u0437\u043E\u0432\
        \u0430\u0442\u0435\u043B\u0435, \u0434\u0430\u043B\u044C\u0448\u0435 \u0438\
        \u0434\u0435\u0442                 \u0437\u0430\u043F\u0438\u0441\u044C \u0432\
        \ \u0411\u0414, \u0432\u044B\u0434\u0430\u0447\u0430 \u0442\u043E\u043A\u0435\
        \u043D\u043E\u0432, \u043A\u043E\u0442\u043E\u0440\u044B\u0435 \u043E\u0442\
        \u0441\u044B\u043B\u0430\u044E\u0442\u0441\u044F                 \u0432 \u043A\
        \u0443\u043A\u0430\u0445 (refresh) \u0438 \u0432 Authorization: Bearer (access)"
      operationId: auth_yandex_issue_jwt_auth_yandex_callback_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Code'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenInfo'
          description: Successful Response
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: "\u041F\u043E\u043B\u044C\u0437\u043E\u0432\u0430\u0442\u0435\
            \u043B\u044C \u0441 \u044D\u0442\u0438 email-\u043E\u043C \u0443\u0436\
            \u0435 \u0441\u0443\u0449\u0435\u0441\u0442\u0432\u0443\u0435\u0442"
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: "\u041D\u0435 \u0432\u0430\u043B\u0438\u0434\u043D\u044B\u0439\
            \ code \u043E\u0442 \u042F\u043D\u0434\u0435\u043A\u0441\u0430"
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: "\u041E\u0442\u043F\u0440\u0430\u0432\u043A\u0430 \u043A\u043E\u0434\
        \u0430, \u043F\u043E\u043B\u0443\u0447\u0435\u043D\u0438\u044F \u0434\u0430\
        \u043D\u043D\u044B\u0445 \u043F\u043E\u043B\u044C\u0437\u043E\u0432\u0430\u0442\
        \u0435\u043B\u044F,                  \u0432\u044B\u0434\u0430\u0447\u0430\
        \ \u0442\u043E\u043A\u0435\u043D\u043E\u0432"
      tags:
      - Auth
...